pre-training pipeline foundation
